+++ waf-openapi-20211001/alibabacloud_waf_openapi20211001/models/_create_defense_rule_request.py
@@ -46,6 +46,10 @@ class CreateDefenseRuleRequest(DaraModel):
         self.resource = resource
         # The ID of the resource group.
         self.resource_manager_resource_group_id = resource_manager_resource_group_id
+        # The configurations of the protection rule. The value is a JSON string that contains multiple parameters.
+        # 
+        # >  The parameters vary based on the **protection module**, which is specified by **DefenseScene**. For more information, see the "**Parameters of protection rules**" section in this topic.
+        # 
         # This parameter is required.
         self.rules = rules
         # The ID of the rule template for which you want to create a protection rule.