+++ waf-openapi-20211001/alibabacloud_waf_openapi20211001/models/_modify_defense_rule_request.py
@@ -33,6 +33,11 @@ class ModifyDefenseRuleRequest(DaraModel):
         self.resource = resource
         # The ID of the Alibaba Cloud resource group.
         self.resource_manager_resource_group_id = resource_manager_resource_group_id
+        # The details of the protection rule. Specify a string that contains multiple parameters in the JSON format. You must specify the ID and the new configurations of the protection rule.
+        # 
+        # *   **id:** The ID of the protection rule. Data type: long. You must specify this parameter.
+        # *   The protection rule configurations: The role of this parameter is the same as that of the **Rules** parameter in the **CreateDefenseRule** topic. For more information, see the "**Protection rule parameters**" section in the [CreateDefenseRule](~~CreateDefenseRule~~) topic.
+        # 
         # This parameter is required.
         self.rules = rules
         # The ID of the protection rule template to which the protection rule whose configurations you want to modify belongs.