+++ waf-openapi-20211001/alibabacloud_waf_openapi20211001/models/_modify_domain_request.py
@@ -124,6 +124,7 @@ class ModifyDomainRequestRedirect(DaraModel):
         keepalive_timeout: int = None,
         loadbalance: str = None,
         max_body_size: int = None,
+        proxy_protocol: bool = None,
         read_timeout: int = None,
         request_headers: List[main_models.ModifyDomainRequestRedirectRequestHeaders] = None,
         retry: bool = None,
@@ -181,6 +182,7 @@ class ModifyDomainRequestRedirect(DaraModel):
         # This parameter is required.
         self.loadbalance = loadbalance
         self.max_body_size = max_body_size
+        self.proxy_protocol = proxy_protocol
         # The timeout period of read connections. Unit: seconds. Valid values: 1 to 3600.
         self.read_timeout = read_timeout
         # The custom header fields, which are key-value pairs. The fields are used to mark requests that pass through WAF.
@@ -275,6 +277,9 @@ class ModifyDomainRequestRedirect(DaraModel):
         if self.max_body_size is not None:
             result['MaxBodySize'] = self.max_body_size
 
+        if self.proxy_protocol is not None:
+            result['ProxyProtocol'] = self.proxy_protocol
+
         if self.read_timeout is not None:
             result['ReadTimeout'] = self.read_timeout
 
@@ -358,6 +363,9 @@ class ModifyDomainRequestRedirect(DaraModel):
 
         if m.get('MaxBodySize') is not None:
             self.max_body_size = m.get('MaxBodySize')
+
+        if m.get('ProxyProtocol') is not None:
+            self.proxy_protocol = m.get('ProxyProtocol')
 
         if m.get('ReadTimeout') is not None:
             self.read_timeout = m.get('ReadTimeout')
@@ -489,6 +497,9 @@ class ModifyDomainRequestListen(DaraModel):
         enable_tlsv_3: bool = None,
         exclusive_ip: bool = None,
         focus_https: bool = None,
+        hsts_include_sub_domain: bool = None,
+        hsts_max_age: int = None,
+        hsts_preload: bool = None,
         http_2enabled: bool = None,
         http_ports: List[int] = None,
         https_ports: List[int] = None,
@@ -526,6 +537,9 @@ class ModifyDomainRequestListen(DaraModel):
         # *   **true**
         # *   **false**
         self.focus_https = focus_https
+        self.hsts_include_sub_domain = hsts_include_sub_domain
+        self.hsts_max_age = hsts_max_age
+        self.hsts_preload = hsts_preload
         # Specifies whether to enable HTTP/2. This parameter is available only if you specify **HttpsPorts**. Valid values:
         # 
         # *   **true**
@@ -597,6 +611,15 @@ class ModifyDomainRequestListen(DaraModel):
         if self.focus_https is not None:
             result['FocusHttps'] = self.focus_https
 
+        if self.hsts_include_sub_domain is not None:
+            result['HstsIncludeSubDomain'] = self.hsts_include_sub_domain
+
+        if self.hsts_max_age is not None:
+            result['HstsMaxAge'] = self.hsts_max_age
+
+        if self.hsts_preload is not None:
+            result['HstsPreload'] = self.hsts_preload
+
         if self.http_2enabled is not None:
             result['Http2Enabled'] = self.http_2enabled
 
@@ -651,6 +674,15 @@ class ModifyDomainRequestListen(DaraModel):
 
         if m.get('FocusHttps') is not None:
             self.focus_https = m.get('FocusHttps')
+
+        if m.get('HstsIncludeSubDomain') is not None:
+            self.hsts_include_sub_domain = m.get('HstsIncludeSubDomain')
+
+        if m.get('HstsMaxAge') is not None:
+            self.hsts_max_age = m.get('HstsMaxAge')
+
+        if m.get('HstsPreload') is not None:
+            self.hsts_preload = m.get('HstsPreload')
 
         if m.get('Http2Enabled') is not None:
             self.http_2enabled = m.get('Http2Enabled')